transport/requirements.txt
Django==6.0.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.4.0
django-cors-headers==4.6.0
mysqlclient==2.2.6
Pillow==11.1.0

channels==4.0.0
daphne==4.0.0
